requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
twilio==8.10.0
openai==1.3.0
pymongo==4.6.0
python-multipart==0.0.6
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
requests==2.31.0
python-dotenv==1.0.0
motor==3.3.1
pytest==7.4.3
httpx==0.25.1 